requirements.txt
flask
flask-cors
numpy
tensorflow-cpu
joblib
